Refactor sharing item, deal with permissions better
